[feat] 调整tools结构
