+++ tools/__init__.py
@@ -4,114 +4,8 @@
   date          : 2024/10/26
   Description   : 工具函数/算子
 """
-import numpy as np
-
-laplace_core = np.array([[ 1,  1,  1],
-                          [ 1, -8,  1],
-                          [ 1,  1,  1]])
-
-sobel_x = np.array([[ -1,  0,  1],
-                     [ -2,  0,  2],
-                     [ -1,  0,  1]])
-
-sobel_y = np.array([[ -1, -2, -1],
-                     [  0,  0,  0],
-                     [  1,  2,  1]])
-
-sobel_diagonal_1 = np.array([[ 0,  1,  2],
-                              [-1,  0,  1],
-                              [-2, -1,  0]])
-
-sobel_diagonal_2 = np.array([[ -2, -1,  0],
-                              [ -1,  0,  1],
-                              [  0,  1,  2]])
-
-sharpen_kernel = np.array([[0, -1, 0],
-                           [-1, 5, -1],
-                           [0, -1, 0]])
-
-def gaussian_kernel(size, sigma):
-    """
-    生成高斯核
-    :param size:
-    :param sigma:
-    :return:
-    """
-    kernel = np.fromfunction(
-        lambda x, y: (1/ (2 * np.pi * sigma**2)) *
-                     np.exp(-((x - (size - 1) / 2)**2 + (y - (size - 1) / 2)**2) / (2 * sigma**2)),
-                    # 在生成高斯核时，我们希望核的中心与图像的当前像素对齐。
-                    # 使用(x-(size-1)/2)可以确保高斯核的中心位于核的中心像素
-        (size, size)
-    )
-    return kernel / np.sum(kernel)  # 归一化
-
-def laplace_of_gaussian_kernel(size, sigma):
-    """
-    生成高斯拉普拉斯核
-    :param size:
-    :param sigma:
-    :return:
-    """
-    kernel = np.fromfunction(
-        lambda x, y: ((1 / sigma ** 4) * (2 * sigma ** 2 - x ** 2 + y ** 2)
-              * -np.exp(x ** 2 + y ** 2 / (2 * sigma ** 2)))
-        (size, size)
-    )
-
-# 卷积
-def conv_with_core(img, knl, pad = False):
-    """
-    卷积运算
-    :param img: array格式的图片
-    :param knl: 卷积核
-    :param pad: 是否填充
-    :return: 卷积后的输出
-    """
-    knl_height, knl_width = knl.shape
-    if pad:
-        pad_height = knl_height // 2
-        pad_width = knl_width // 2
-        img = np.pad(img, ((pad_height, pad_height), (pad_width, pad_width)), mode='edge')
-
-    img_height, img_width = img.shape
-    ret_height = img_height - knl_height + 1
-    ret_width = img_width - knl_width + 1
-    ret = np.zeros((ret_height, ret_width))
-    for i in range(ret_height):
-        for j in range(ret_width):
-            rgn = img[i:i+knl_height, j:j+knl_width]
-            ret[i, j] = np.sum(rgn * knl)
-    return ret
-
-def median_filter(img, knl_size):
-    """
-    对图像应用中值滤波
-    :param img:
-    :param knl_size:
-    :return:
-    """
-    if knl_size % 2 == 0:
-        raise ValueError("knl_size must be odd number")
-    pad_size = knl_size // 2
-    pad_img = np.pad(img, pad_size, mode='edge')
-    ret = np.zeros_like(img)
-    for i in range(img.shape[0]):
-        for j in range(img.shape[1]):
-            # 提取当前像素及其邻域
-            rgn = pad_img[i:i+knl_size, j:j+knl_size]
-            ret[i, j] = np.median(rgn)
-    return ret
-
-# 激活函数
-def sigmoid(x):
-    return 1 / (1 + np.exp(-x))
-def relu(x):
-    return np.maximum(0, x)
-def tanh(x):
-    return np.tanh(x)
-# 阈值函数
-def apply_threshold(img, threshold):
-    ret = np.zeros_like(img)
-    ret[img > threshold] = img[img > threshold]
-    return ret
+from .kernels import laplace_core, sobel_x, sobel_y, sobel_diagonal_1, sobel_diagonal_2
+from .kernels import sharpen_kernel, gaussian_kernel, laplace_of_gaussian_kernel
+from .functions import conv_with_core, median_filter
+from .functions import sigmoid, relu, tanh
+from .functions import apply_threshold
